requirements.txt
Flask==3.0.2
twilio==9.2.3
python-dotenv==1.0.1
openai==1.14.2
requests==2.31.0